[fix] request 유효성 검사 수정
추가로 자바 컨벤션에 맞게 코드 재정렬

+++ src/main/java/shop/bookbom/shop/domain/wish/controller/WishController.java
@@ -1,6 +1,7 @@
 package shop.bookbom.shop.domain.wish.controller;
 
 import java.util.List;
+import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -11,9 +12,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import shop.bookbom.shop.common.CommonListResponse;
 import shop.bookbom.shop.common.CommonResponse;
+import shop.bookbom.shop.domain.cart.dto.request.CartAddRequest;
 import shop.bookbom.shop.domain.wish.dto.request.WishAddDeleteRequest;
 import shop.bookbom.shop.domain.wish.dto.response.WishInfoResponse;
 import shop.bookbom.shop.domain.wish.dto.response.WishTotalCountResponse;
+import shop.bookbom.shop.domain.wish.exception.WishInvalidRequestException;
 import shop.bookbom.shop.domain.wish.service.WishService;
 
 @RestController
@@ -23,7 +26,6 @@ public class WishController {
     private final WishService wishService;
 
     /**
-     *
      * 찜 목록에 도서를 추가합니다
      *
      * @param userId
@@ -31,13 +33,16 @@ public class WishController {
      * @return
      */
     @PostMapping("/wish/{id}")
-    public CommonResponse<Void> addWish(@PathVariable("id") Long userId, @RequestBody List<WishAddDeleteRequest> request) {
+    public CommonResponse<Void> addWish(@PathVariable("id") @Valid Long userId,
+                                        @RequestBody List<WishAddDeleteRequest> request) {
+        if (!isValidRequest(request)) {
+            throw new WishInvalidRequestException();
+        }
         wishService.addWish(request, userId);
         return CommonResponse.success();
     }
 
     /**
-     *
      * 찜 목록에 있는 도서를 삭제합니다
      *
      * @param userId
@@ -45,13 +50,16 @@ public class WishController {
      * @return
      */
     @DeleteMapping("/wish/{id}")
-    public CommonResponse<Void> deleteWish(@PathVariable("id") Long userId, @RequestBody List<WishAddDeleteRequest> request) {
+    public CommonResponse<Void> deleteWish(@PathVariable("id") Long userId,
+                                           @RequestBody List<WishAddDeleteRequest> request) {
+        if (!isValidRequest(request)) {
+            throw new WishInvalidRequestException();
+        }
         wishService.deleteWish(request, userId);
         return CommonResponse.success();
     }
 
     /**
-     *
      * 회원의 찜 목록을 조회합니다
      *
      * @param userId
@@ -63,14 +71,25 @@ public class WishController {
     }
 
     /**
-     *
      * 회원의 모든 찜 개수를 조회합니다
      *
      * @param userId
      * @return
      */
     @GetMapping("/wish/count/{id}")
-    public CommonResponse<WishTotalCountResponse> getWishCount(@PathVariable("id") Long userId){
+    public CommonResponse<WishTotalCountResponse> getWishCount(@PathVariable("id") Long userId) {
         return CommonResponse.successWithData(wishService.getWishTotalCount(userId));
+    }
+
+    /**
+     * request 유효성 검사
+     *
+     * @param requests
+     * @return
+     */
+    private boolean isValidRequest(List<WishAddDeleteRequest> requests) {
+        return requests.stream()
+                .allMatch(request -> request.getBookId() != null
+                        && request.getBookId() >= 1);
     }
 }
